Extract stage collision loader
